Rename CodeCraft to ScotRUG in the Code of Conduct
The reference to “CodeCraft” was added along with the Code of Conduct when
it was copied from [their website][1] in 48c885aa7.  I’ve changed it to
“ScotRUG”, assuming it was an oversight.

I’ve also changed the capitalization of “ScotRug” to “ScotRUG” across
the Code of Conduct for consistency with the rest of the site.

[1]: http://www.codecraftuk.org/code-of-conduct.html