Comment out iframes on index page
Several posts include iframes to load in videos from Vimeo.  Each iframe
requests CSS, JS and images, contributing significantly to the size of
the home page, which currently indexes five years of posts.

This commit wraps iframes on the home page in comments (full removal
would be trickier with Liquid, as regular expressions are not supported
by default.)

This reduces the number of requests made when visiting scotrug.org from
100 to 12, and reduces cold cache page weight from 6,125 KB to 340 KB.
On my machine this results in a page load time from localhost of less
than 5 percent of the original time.

Each video is captioned so their links remain on the home page.  Videos
also remain embedded in individual posts.