Adds pagination to the site
The blog has been a one-page thing since it was built. Though it builds
quickly, there's quite a lot of information to page through and makes
the page feel odd. The upside to this is that the page will load
slightly faster (though not much faster).

The downside is that it becomes harder to find historical posts, as you
can't user your browser to find in page.

I think overall it works better. We've chosen 12 posts per page, to
reflect a calendar year, though note that we rarely run 12 meetups a
year.