Updates Ruby version and dependencies
Github pages has been in a bit of quiet space for some time. Late last
year, there was an update. It's still using Jekyll 3, but has now been
updated to Ruby 3.

This keeps us up to date, and gives us the change to move and update the
site in the near future.